Added Colors and shape change, has color bugs and unconcsistent popping :( trying to respawn shapes using transform.

+++ Assets/Scripts/Tower.cs
@@ -8,7 +8,7 @@ public class Tower : MonoBehaviour
     public int damage;
     public int reach;
     public int bulletSpeed;
-    public int reload;
+    public float reload;
     public List<Enemy> oppReach;
     private CircleCollider2D col;
     public GameObject projectilePrefab;
